+++ DATA/Models/Convenios.cs
@@ -12,7 +12,7 @@ namespace DATA.Models
         }
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public long IdConvenio { get; set; }//Entidad base
+        public int IdConvenio { get; set; }//Entidad base
         [MaxLength(100)]
         public string Descripcion { get; set; }
         public string Obs { get; set; } //Entidad base
